@@ -1,5 +1,5 @@
-//go:build !ignore_autogenerated_conversions
-// +build !ignore_autogenerated_conversions
+//go:build !ignore_autogenerated
+// +build !ignore_autogenerated
 
 /*
 Copyright The Kubernetes Authors.
@@ -2082,6 +2082,7 @@ func autoConvert_v1beta2_NetworkSpec_To_v1beta1_NetworkSpec(in *v1beta2.NetworkS
 	out.CNI = (*CNISpec)(unsafe.Pointer(in.CNI))
 	out.SecurityGroupOverrides = *(*map[SecurityGroupRole]string)(unsafe.Pointer(&in.SecurityGroupOverrides))
 	// WARNING: in.AdditionalControlPlaneIngressRules requires manual conversion: does not exist in peer-type
+	// WARNING: in.NodePortIngressRuleCidrBlocks requires manual conversion: does not exist in peer-type
 	return nil
 }
 
@@ -2300,6 +2301,7 @@ func Convert_v1beta1_VPCSpec_To_v1beta2_VPCSpec(in *VPCSpec, out *v1beta2.VPCSpe
 func autoConvert_v1beta2_VPCSpec_To_v1beta1_VPCSpec(in *v1beta2.VPCSpec, out *VPCSpec, s conversion.Scope) error {
 	out.ID = in.ID
 	out.CidrBlock = in.CidrBlock
+	// WARNING: in.SecondaryCidrBlocks requires manual conversion: does not exist in peer-type
 	// WARNING: in.IPAMPool requires manual conversion: does not exist in peer-type
 	if in.IPv6 != nil {
 		in, out := &in.IPv6, &out.IPv6
