@@ -1,5 +1,5 @@
-//go:build !ignore_autogenerated_conversions
-// +build !ignore_autogenerated_conversions
+//go:build !ignore_autogenerated
+// +build !ignore_autogenerated
 
 /*
 Copyright The Kubernetes Authors.
