@@ -1,5 +1,5 @@
-//go:build !ignore_autogenerated_conversions
-// +build !ignore_autogenerated_conversions
+//go:build !ignore_autogenerated
+// +build !ignore_autogenerated
 
 /*
 Copyright The Kubernetes Authors.
@@ -310,6 +310,11 @@ func RegisterConversions(s *runtime.Scheme) error {
 	}); err != nil {
 		return err
 	}
+	if err := s.AddConversionFunc((*apiv1beta2.Instance)(nil), (*apiv1beta1.Instance)(nil), func(a, b interface{}, scope conversion.Scope) error {
+		return Convert_v1beta2_Instance_To_v1beta1_Instance(a.(*apiv1beta2.Instance), b.(*apiv1beta1.Instance), scope)
+	}); err != nil {
+		return err
+	}
 	if err := s.AddConversionFunc((*v1beta2.RefreshPreferences)(nil), (*RefreshPreferences)(nil), func(a, b interface{}, scope conversion.Scope) error {
 		return Convert_v1beta2_RefreshPreferences_To_v1beta1_RefreshPreferences(a.(*v1beta2.RefreshPreferences), b.(*RefreshPreferences), scope)
 	}); err != nil {
@@ -402,6 +407,7 @@ func autoConvert_v1beta2_AWSLaunchTemplate_To_v1beta1_AWSLaunchTemplate(in *v1be
 	out.ImageLookupBaseOS = in.ImageLookupBaseOS
 	out.InstanceType = in.InstanceType
 	out.RootVolume = (*apiv1beta2.Volume)(unsafe.Pointer(in.RootVolume))
+	// WARNING: in.NonRootVolumes requires manual conversion: does not exist in peer-type
 	out.SSHKeyName = (*string)(unsafe.Pointer(in.SSHKeyName))
 	out.VersionNumber = (*int64)(unsafe.Pointer(in.VersionNumber))
 	out.AdditionalSecurityGroups = *(*[]apiv1beta2.AWSResourceReference)(unsafe.Pointer(&in.AdditionalSecurityGroups))
@@ -1065,6 +1071,7 @@ func autoConvert_v1beta2_RefreshPreferences_To_v1beta1_RefreshPreferences(in *v1
 	out.Strategy = (*string)(unsafe.Pointer(in.Strategy))
 	out.InstanceWarmup = (*int64)(unsafe.Pointer(in.InstanceWarmup))
 	out.MinHealthyPercentage = (*int64)(unsafe.Pointer(in.MinHealthyPercentage))
+	// WARNING: in.MaxHealthyPercentage requires manual conversion: does not exist in peer-type
 	return nil
 }
 
