@@ -24,10 +24,13 @@ import (
 	"fmt"
 	"net"
 	"net/url"
+	"reflect"
 	"strconv"
 	"strings"
 	"time"
 
+	stsv2 "github.com/aws/aws-sdk-go-v2/service/sts"
+	sts "github.com/aws/aws-sdk-go/service/sts"
 	"github.com/google/go-cmp/cmp"
 	idputils "github.com/openshift-online/ocm-common/pkg/idp/utils"
 	cmv1 "github.com/openshift-online/ocm-sdk-go/clustersmgmt/v1"
@@ -104,16 +107,16 @@ func (r *ROSAControlPlaneReconciler) SetupWithManager(ctx context.Context, mgr c
 	}
 
 	if err = c.Watch(
-		source.Kind(mgr.GetCache(), &clusterv1.Cluster{}),
-		handler.EnqueueRequestsFromMapFunc(util.ClusterToInfrastructureMapFunc(ctx, rosaControlPlane.GroupVersionKind(), mgr.GetClient(), &expinfrav1.ROSACluster{})),
-		predicates.ClusterUnpausedAndInfrastructureReady(log.GetLogger()),
+		source.Kind[client.Object](mgr.GetCache(), &clusterv1.Cluster{},
+			handler.EnqueueRequestsFromMapFunc(util.ClusterToInfrastructureMapFunc(ctx, rosaControlPlane.GroupVersionKind(), mgr.GetClient(), &expinfrav1.ROSACluster{})),
+			predicates.ClusterUnpausedAndInfrastructureReady(log.GetLogger())),
 	); err != nil {
 		return fmt.Errorf("failed adding a watch for ready clusters: %w", err)
 	}
 
 	if err = c.Watch(
-		source.Kind(mgr.GetCache(), &expinfrav1.ROSACluster{}),
-		handler.EnqueueRequestsFromMapFunc(r.rosaClusterToROSAControlPlane(log)),
+		source.Kind[client.Object](mgr.GetCache(), &expinfrav1.ROSACluster{},
+			handler.EnqueueRequestsFromMapFunc(r.rosaClusterToROSAControlPlane(log))),
 	); err != nil {
 		return fmt.Errorf("failed adding a watch for ROSACluster")
 	}
@@ -206,7 +209,7 @@ func (r *ROSAControlPlaneReconciler) reconcileNormal(ctx context.Context, rosaSc
 		return ctrl.Result{}, fmt.Errorf("failed to create OCM client: %w", err)
 	}
 
-	creator, err := rosaaws.CreatorForCallerIdentity(rosaScope.Identity)
+	creator, err := rosaaws.CreatorForCallerIdentity(convertStsV2(rosaScope.Identity))
 	if err != nil {
 		return ctrl.Result{}, fmt.Errorf("failed to transform caller identity to creator: %w", err)
 	}
@@ -334,7 +337,7 @@ func (r *ROSAControlPlaneReconciler) reconcileDelete(ctx context.Context, rosaSc
 		return ctrl.Result{}, fmt.Errorf("failed to create OCM client: %w", err)
 	}
 
-	creator, err := rosaaws.CreatorForCallerIdentity(rosaScope.Identity)
+	creator, err := rosaaws.CreatorForCallerIdentity(convertStsV2(rosaScope.Identity))
 	if err != nil {
 		return ctrl.Result{}, fmt.Errorf("failed to transform caller identity to creator: %w", err)
 	}
@@ -439,26 +442,88 @@ func (r *ROSAControlPlaneReconciler) reconcileClusterVersion(rosaScope *scope.RO
 }
 
 func (r *ROSAControlPlaneReconciler) updateOCMCluster(rosaScope *scope.ROSAControlPlaneScope, ocmClient *ocm.Client, cluster *cmv1.Cluster, creator *rosaaws.Creator) error {
-	currentAuditLogRole := cluster.AWS().AuditLog().RoleArn()
-	if currentAuditLogRole == rosaScope.ControlPlane.Spec.AuditLogRoleARN {
-		return nil
-	}
+	ocmClusterSpec, updated := r.updateOCMClusterSpec(rosaScope.ControlPlane, cluster)
 
-	ocmClusterSpec := ocm.Spec{
-		AuditLogRoleARN: ptr.To(rosaScope.ControlPlane.Spec.AuditLogRoleARN),
-	}
-
-	// if this fails, the provided role is likely invalid or it doesn't have the required permissions.
-	if err := ocmClient.UpdateCluster(cluster.ID(), creator, ocmClusterSpec); err != nil {
-		conditions.MarkFalse(rosaScope.ControlPlane,
-			rosacontrolplanev1.ROSAControlPlaneValidCondition,
-			rosacontrolplanev1.ROSAControlPlaneInvalidConfigurationReason,
-			clusterv1.ConditionSeverityError,
-			err.Error())
-		return err
+	if updated {
+		// Update the cluster.
+		rosaScope.Info("Updating cluster")
+		if err := ocmClient.UpdateCluster(cluster.ID(), creator, ocmClusterSpec); err != nil {
+			conditions.MarkFalse(rosaScope.ControlPlane,
+				rosacontrolplanev1.ROSAControlPlaneValidCondition,
+				rosacontrolplanev1.ROSAControlPlaneInvalidConfigurationReason,
+				clusterv1.ConditionSeverityError,
+				err.Error())
+			return err
+		}
 	}
 
 	return nil
+}
+
+func (r *ROSAControlPlaneReconciler) updateOCMClusterSpec(rosaControlPlane *rosacontrolplanev1.ROSAControlPlane, cluster *cmv1.Cluster) (ocm.Spec, bool) {
+	ocmClusterSpec := ocm.Spec{}
+	updated := false
+
+	// Check for audit role arn changes
+	currentAuditLogRole := cluster.AWS().AuditLog().RoleArn()
+	if currentAuditLogRole != rosaControlPlane.Spec.AuditLogRoleARN {
+		ocmClusterSpec.AuditLogRoleARN = ptr.To(rosaControlPlane.Spec.AuditLogRoleARN)
+		updated = true
+	}
+
+	// Check for registry config changes
+	regConfig := &rosacontrolplanev1.RegistryConfig{
+		RegistrySources: &rosacontrolplanev1.RegistrySources{},
+	}
+	if rosaControlPlane.Spec.ClusterRegistryConfig != nil {
+		regConfig.AdditionalTrustedCAs = rosaControlPlane.Spec.ClusterRegistryConfig.AdditionalTrustedCAs
+		regConfig.AllowedRegistriesForImport = rosaControlPlane.Spec.ClusterRegistryConfig.AllowedRegistriesForImport
+
+		if rosaControlPlane.Spec.ClusterRegistryConfig.RegistrySources != nil {
+			regConfig.RegistrySources.AllowedRegistries = rosaControlPlane.Spec.ClusterRegistryConfig.RegistrySources.AllowedRegistries
+			regConfig.RegistrySources.BlockedRegistries = rosaControlPlane.Spec.ClusterRegistryConfig.RegistrySources.BlockedRegistries
+			regConfig.RegistrySources.InsecureRegistries = rosaControlPlane.Spec.ClusterRegistryConfig.RegistrySources.InsecureRegistries
+		}
+	}
+	if !reflect.DeepEqual(regConfig.AdditionalTrustedCAs, cluster.RegistryConfig().AdditionalTrustedCa()) {
+		ocmClusterSpec.AdditionalTrustedCa = regConfig.AdditionalTrustedCAs
+		updated = true
+	}
+	if !reflect.DeepEqual(regConfig.RegistrySources.AllowedRegistries, cluster.RegistryConfig().RegistrySources().AllowedRegistries()) {
+		ocmClusterSpec.AllowedRegistries = regConfig.RegistrySources.AllowedRegistries
+		updated = true
+	}
+	if !reflect.DeepEqual(regConfig.RegistrySources.BlockedRegistries, cluster.RegistryConfig().RegistrySources().BlockedRegistries()) {
+		ocmClusterSpec.BlockedRegistries = regConfig.RegistrySources.BlockedRegistries
+		updated = true
+	}
+	if !reflect.DeepEqual(regConfig.RegistrySources.InsecureRegistries, cluster.RegistryConfig().RegistrySources().InsecureRegistries()) {
+		ocmClusterSpec.InsecureRegistries = regConfig.RegistrySources.InsecureRegistries
+		updated = true
+	}
+
+	var newAllowedRegistries, oldAllowedRegistries []string
+	if len(regConfig.AllowedRegistriesForImport) > 0 {
+		for id := range regConfig.AllowedRegistriesForImport {
+			newAllowedRegistries = append(newAllowedRegistries, regConfig.AllowedRegistriesForImport[id].DomainName+":"+
+				strconv.FormatBool(regConfig.AllowedRegistriesForImport[id].Insecure))
+		}
+	}
+	if len(cluster.RegistryConfig().AllowedRegistriesForImport()) > 0 {
+		for id := range cluster.RegistryConfig().AllowedRegistriesForImport() {
+			oldAllowedRegistries = append(oldAllowedRegistries, cluster.RegistryConfig().AllowedRegistriesForImport()[id].DomainName()+":"+
+				strconv.FormatBool(cluster.RegistryConfig().AllowedRegistriesForImport()[id].Insecure()))
+		}
+	}
+	if !reflect.DeepEqual(newAllowedRegistries, oldAllowedRegistries) {
+		ocmClusterSpec.AllowedRegistriesForImport = strings.Join(newAllowedRegistries, ",")
+		updated = true
+	}
+
+	// TODO: check for cluster AutoScale changes
+	// rosaControlPlane.Spec.DefaultMachinePoolSpec.Autoscaling
+
+	return ocmClusterSpec, updated
 }
 
 func (r *ROSAControlPlaneReconciler) reconcileExternalAuth(ctx context.Context, rosaScope *scope.ROSAControlPlaneScope, cluster *cmv1.Cluster) error {
@@ -888,6 +953,28 @@ func buildOCMClusterSpec(controlPlaneSpec rosacontrolplanev1.RosaControlPlaneSpe
 		}
 	}
 
+	// Set the cluster registry config.
+	if controlPlaneSpec.ClusterRegistryConfig != nil {
+		if len(controlPlaneSpec.ClusterRegistryConfig.AdditionalTrustedCAs) > 0 {
+			ocmClusterSpec.AdditionalTrustedCa = controlPlaneSpec.ClusterRegistryConfig.AdditionalTrustedCAs
+		}
+
+		if len(controlPlaneSpec.ClusterRegistryConfig.AllowedRegistriesForImport) > 0 {
+			registers := make([]string, 0)
+			for id := range controlPlaneSpec.ClusterRegistryConfig.AllowedRegistriesForImport {
+				registers = append(registers, controlPlaneSpec.ClusterRegistryConfig.AllowedRegistriesForImport[id].DomainName+":"+
+					strconv.FormatBool(controlPlaneSpec.ClusterRegistryConfig.AllowedRegistriesForImport[id].Insecure))
+			}
+			ocmClusterSpec.AllowedRegistriesForImport = strings.Join(registers, ",")
+		}
+
+		if controlPlaneSpec.ClusterRegistryConfig.RegistrySources != nil {
+			ocmClusterSpec.BlockedRegistries = controlPlaneSpec.ClusterRegistryConfig.RegistrySources.BlockedRegistries
+			ocmClusterSpec.AllowedRegistries = controlPlaneSpec.ClusterRegistryConfig.RegistrySources.AllowedRegistries
+			ocmClusterSpec.InsecureRegistries = controlPlaneSpec.ClusterRegistryConfig.RegistrySources.InsecureRegistries
+		}
+	}
+
 	return ocmClusterSpec, nil
 }
 
@@ -995,4 +1082,13 @@ func buildAPIEndpoint(cluster *cmv1.Cluster) (*clusterv1.APIEndpoint, error) {
 		Host: host,
 		Port: int32(port), // #nosec G109
 	}, nil
+}
+
+// TODO: Remove this and update the aws-sdk lib to v2.
+func convertStsV2(identity *sts.GetCallerIdentityOutput) *stsv2.GetCallerIdentityOutput {
+	return &stsv2.GetCallerIdentityOutput{
+		Account: identity.Account,
+		Arn:     identity.Arn,
+		UserId:  identity.UserId,
+	}
 }
