@@ -100,7 +100,7 @@ func (p *defaultSubnetPlacementStrategy) Place(input *placementInput) ([]string,
 		return subnetIDs, nil
 	}
 
-	controlPlaneSubnetIDs := input.ControlplaneSubnets.FilterPrivate().IDs()
+	controlPlaneSubnetIDs := input.ControlplaneSubnets.FilterPrivate().FilterNonCni().IDs()
 	if len(controlPlaneSubnetIDs) > 0 {
 		p.logger.Debug("using all the private subnets from the control plane")
 		return controlPlaneSubnetIDs, nil
@@ -119,10 +119,9 @@ func (p *defaultSubnetPlacementStrategy) getSubnetsForAZs(azs []string, controlP
 			case expinfrav1.AZSubnetTypeAll:
 				// no-op
 			case expinfrav1.AZSubnetTypePublic:
-				subnets = subnets.FilterPublic()
+				subnets = subnets.FilterPublic().FilterNonCni()
 			case expinfrav1.AZSubnetTypePrivate:
-				subnets = subnets.FilterPrivate()
-				subnets = subnets.FilterNonCni()
+				subnets = subnets.FilterPrivate().FilterNonCni()
 			}
 		}
 		if len(subnets) == 0 {
