@@ -117,6 +117,7 @@ var _ = ginkgo.Context("[unmanaged] [functional]", func() {
 
 	ginkgo.Describe("GPU-enabled cluster test", func() {
 		ginkgo.It("should create cluster with single worker", func() {
+			ginkgo.Skip("Args field of clusterctl.ApplyClusterTemplateAndWaitInput was removed, need to add support for server-side filtering.")
 			specName := "functional-gpu-cluster"
 			// Change the multiplier for EC2GPU if GPU type is changed. g4dn.xlarge uses 2 vCPU
 			requiredResources = &shared.TestResource{EC2GPU: 2 * 2, IGW: 1, NGW: 1, VPC: 1, ClassicLB: 1, EIP: 1, EventBridgeRules: 50}
@@ -149,7 +150,9 @@ var _ = ginkgo.Context("[unmanaged] [functional]", func() {
 				// This is because the entire config map is stored in `last-applied` annotation for tracking.
 				// The workaround is to use server side apply by passing `--server-side` flag to kubectl apply.
 				// More on server side apply here: https://kubernetes.io/docs/reference/using-api/server-side-apply/
-				Args: []string{"--server-side"},
+				// TODO: Need a PR to re-add argument support to this type.
+				// It was removed in https://github.com/kubernetes-sigs/cluster-api/commit/b4349fecaa626865e71b058a8b01e0377fb9e444
+				// Args: []string{"--server-side"},
 			}, result)
 
 			shared.AWSGPUSpec(ctx, e2eCtx, shared.AWSGPUSpecInput{
@@ -1009,7 +1012,7 @@ var _ = ginkgo.Context("[unmanaged] [functional]", func() {
 				})
 
 				framework.WaitForClusterDeleted(ctx, framework.WaitForClusterDeletedInput{
-					Getter:  mgmtClusterProxy.GetClient(),
+					Client:  mgmtClusterProxy.GetClient(),
 					Cluster: wlResult.Cluster,
 				}, e2eCtx.E2EConfig.GetIntervals("", "wait-delete-cluster")...)
 
